+++ packages/stryker-karma-runner/test/integration/KarmaTestRunner.it.ts
@@ -1,8 +1,10 @@
 import { expect } from 'chai';
+import * as http from 'http';
 import { CoverageCollection, RunnerOptions, RunResult, RunStatus, TestStatus } from 'stryker-api/test_runner';
 import KarmaTestRunner from '../../src/KarmaTestRunner';
 import JasmineTestFramework from 'stryker-jasmine/src/JasmineTestFramework';
 import { expectTestResults } from '../helpers/assertions';
+import { promisify } from '@stryker-mutator/util';
 
 function wrapInClosure(codeFragment: string) {
   return `
@@ -238,4 +240,73 @@ describe('KarmaTestRunner', () => {
       return true;
     }));
   });
+
+  describe('when specified port is not available', () => {
+
+    let dummyServer: DummyServer;
+
+    before(async () => {
+      dummyServer = await DummyServer.create();
+
+      const testRunnerOptions: RunnerOptions = {
+        fileNames: [],
+        port: dummyServer.port,
+        strykerOptions: {
+          karma: {
+            config: {
+              files: [
+                'testResources/sampleProject/src-instrumented/Add.js',
+                'testResources/sampleProject/test/AddSpec.js'
+              ]
+            }
+          }
+        }
+      };
+      sut = new KarmaTestRunner(testRunnerOptions);
+      return sut.init();
+    });
+
+    after(async () => {
+      if (dummyServer) {
+        await dummyServer.dispose();
+      }
+    });
+
+    it('should choose different port automatically and report Complete without errors', async () => {
+      const actualResult = await sut.run({});
+      expect(sut.port).not.eq(dummyServer.port);
+      expect(actualResult.status, JSON.stringify(actualResult.errorMessages)).eq(RunStatus.Complete);
+    });
+  });
 });
+
+class DummyServer {
+  private readonly httpServer: http.Server;
+
+  private constructor() {
+    this.httpServer = http.createServer();
+  }
+
+  get port() {
+    const address = this.httpServer.address();
+    if (typeof address === 'string') {
+      throw new Error(`Address "${address}" was unexpected: https://nodejs.org/dist/latest-v11.x/docs/api/net.html#net_server_address`);
+    } else {
+      return address.port;
+    }
+  }
+
+  public static async create(): Promise<DummyServer> {
+    const server = new DummyServer();
+    await server.init();
+    return server;
+  }
+
+  private async init(): Promise<void> {
+    await promisify(this.httpServer.listen.bind(this.httpServer))(0, '0.0.0.0');
+  }
+
+  public dispose(): Promise<void> {
+    return promisify(this.httpServer.close.bind(this.httpServer))();
+  }
+}