+++ packages/stryker-webpack-transpiler/src/fs/InputFileSystem.ts
@@ -2,14 +2,18 @@ import MemoryFS from './MemoryFS';
 import { webpack, Callback } from '../types';
 import * as fs from 'fs';
 import { dirname } from 'path';
-import { CachedInputFileSystem, NodeJsInputFileSystem, Stats } from 'enhanced-resolve';
+import {
+  CachedInputFileSystem,
+  NodeJsInputFileSystem,
+  Stats
+} from 'enhanced-resolve';
 
 // Cache duration is same as webpack has
 // => https://github.com/webpack/webpack/blob/efc576c8b744e7a015ab26f1f46932ba3ca7d4f1/lib/node/NodeEnvironmentPlugin.js#L14
 const CACHE_DURATION = 60000;
 
-export default class InputFileSystem extends CachedInputFileSystem implements webpack.InputFileSystem {
-
+export default class InputFileSystem extends CachedInputFileSystem
+  implements webpack.InputFileSystem {
   private readonly memoryFS = new MemoryFS();
 
   constructor(innerFS = new NodeJsInputFileSystem()) {
@@ -26,11 +30,10 @@ export default class InputFileSystem extends CachedInputFileSystem implements we
   }
 
   public stat(path: string, callback: Callback<fs.Stats>): void {
-    this.memoryFS.stat(path, (err: Error, stats: any) => {
+    this.memoryFS.stat(path, (err?: Error | null, stats?: any) => {
       if (err) {
-        super.stat(path, callback);
-      }
-      else {
+        super.stat(path, callback as Callback<Stats>);
+      } else {
         callback(err, stats);
       }
     });