+++ packages/stryker/src/initializer/StrykerInitializer.ts
@@ -6,6 +6,8 @@ import { getLogger } from 'stryker-api/logging';
 import { filterEmpty } from '../utils/objectUtils';
 import StrykerConfigWriter from './StrykerConfigWriter';
 import CommandTestRunner from '../test-runner/CommandTestRunner';
+import StrykerPresets from './StrykerPresets';
+import Preset from './presets/Preset';
 
 const enum PackageManager {
   Npm = 'npm',
@@ -17,7 +19,7 @@ export default class StrykerInitializer {
   private readonly log = getLogger(StrykerInitializer.name);
   private readonly inquirer = new StrykerInquirer();
 
-  constructor(private readonly out = console.log, private readonly client: NpmClient = new NpmClient()) { }
+  constructor(private readonly out = console.log, private readonly client: NpmClient = new NpmClient(), private readonly strykerPresets: Preset[] = StrykerPresets) { }
 
   /**
    * Runs the initializer will prompt the user for questions about his setup. After that, install plugins and configure Stryker.
@@ -27,6 +29,50 @@ export default class StrykerInitializer {
     const configWriter = new StrykerConfigWriter(this.out);
     configWriter.guardForExistingConfig();
     this.patchProxies();
+    const selectedPreset = await this.selectPreset();
+    if (selectedPreset) {
+      await this.initiatePreset(configWriter, selectedPreset);
+    }
+    else {
+      await this.initiateCustom(configWriter);
+    }
+    this.out('Done configuring stryker. Please review `stryker.conf.js`, you might need to configure transpilers or your test runner correctly.');
+    this.out('Let\'s kill some mutants with this command: `stryker run`');
+  }
+
+  /**
+   * The typed rest client works only with the specific HTTP_PROXY and HTTPS_PROXY env settings.
+   * Let's make sure they are available.
+   */
+  private patchProxies() {
+    const copyEnvVariable = (from: string, to: string) => {
+      if (process.env[from] && !process.env[to]) {
+        process.env[to] = process.env[from];
+      }
+    };
+    copyEnvVariable('http_proxy', 'HTTP_PROXY');
+    copyEnvVariable('https_proxy', 'HTTPS_PROXY');
+  }
+
+  private async selectPreset(): Promise<Preset | undefined> {
+    const presetOptions: Preset[] = this.strykerPresets;
+    if (presetOptions.length) {
+      this.log.debug(`Found presets: ${JSON.stringify(presetOptions)}`);
+      return this.inquirer.promptPresets(presetOptions);
+    } else {
+      this.log.debug('No presets have been configured, reverting to custom configuration');
+      return undefined;
+    }
+  }
+
+  private async initiatePreset(configWriter: StrykerConfigWriter, selectedPreset: Preset) {
+    const presetConfig = await selectedPreset.createConfig();
+    await configWriter.writePreset(presetConfig);
+    const selectedPackageManager = await this.selectPackageManager();
+    this.installNpmDependencies(presetConfig.dependencies, selectedPackageManager);
+  }
+
+  private async initiateCustom(configWriter: StrykerConfigWriter) {
     const selectedTestRunner = await this.selectTestRunner();
     const selectedTestFramework = selectedTestRunner && !CommandTestRunner.is(selectedTestRunner.name)
       ? await this.selectTestFramework(selectedTestRunner) : null;
@@ -47,22 +93,6 @@ export default class StrykerInitializer {
       selectedPackageManager,
       await this.fetchAdditionalConfig(npmDependencies));
     this.installNpmDependencies(npmDependencies, selectedPackageManager);
-    this.out('Done configuring stryker. Please review `stryker.conf.js`, you might need to configure transpilers or your test runner correctly.');
-    this.out('Let\'s kill some mutants with this command: `stryker run`');
-  }
-
-  /**
-   * The typed rest client works only with the specific HTTP_PROXY and HTTPS_PROXY env settings.
-   * Let's make sure they are available.
-   */
-  private patchProxies() {
-    const copyEnvVariable = (from: string, to: string) => {
-      if (process.env[from] && !process.env[to]) {
-        process.env[to] = process.env[from];
-      }
-    };
-    copyEnvVariable('http_proxy', 'HTTP_PROXY');
-    copyEnvVariable('https_proxy', 'HTTPS_PROXY');
   }
 
   private async selectTestRunner(): Promise<PromptOption | null> {