+++ packages/stryker/src/initializer/presets/PresetConfiguration.ts
@@ -1,0 +1,5 @@
+export default interface PresetConfiguration {
+  config: string;
+  handbookUrl: string;
+  dependencies: string[];
+}