+++ packages/stryker-javascript-mutator/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.1"></a>
+## [0.12.1](https://github.com/stryker-mutator/stryker/compare/stryker-javascript-mutator@0.12.0...stryker-javascript-mutator@0.12.1) (2018-11-29)
+
+
+### Bug Fixes
+
+* **stryker-api:** Update stryker-api peer dependency version ([677fc28](https://github.com/stryker-mutator/stryker/commit/677fc28))
+
+
+
+
+<a name="0.12.0"></a>
+# [0.12.0](https://github.com/stryker-mutator/stryker/compare/stryker-javascript-mutator@0.11.5...stryker-javascript-mutator@0.12.0) (2018-11-29)
+
+
+### Bug Fixes
+
+* **String literal mutator:** Don't mutate export declarations ([c764ccd](https://github.com/stryker-mutator/stryker/commit/c764ccd))
+
+
+### Features
+
+* **Conditional expression mutator:** Mutate conditional operators ([#1253](https://github.com/stryker-mutator/stryker/issues/1253)) ([be4c990](https://github.com/stryker-mutator/stryker/commit/be4c990))
+
+
+
+
+<a name="0.11.5"></a>
+## [0.11.5](https://github.com/stryker-mutator/stryker/compare/stryker-javascript-mutator@0.11.4...stryker-javascript-mutator@0.11.5) (2018-11-21)
+
+
+
+
+**Note:** Version bump only for package stryker-javascript-mutator
+
 <a name="0.11.4"></a>
 ## [0.11.4](https://github.com/stryker-mutator/stryker/compare/stryker-javascript-mutator@0.11.3...stryker-javascript-mutator@0.11.4) (2018-11-13)