+++ packages/stryker-mocha-runner/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.15.2"></a>
+## [0.15.2](https://github.com/stryker-mutator/stryker/compare/stryker-mocha-runner@0.15.1...stryker-mocha-runner@0.15.2) (2018-11-29)
+
+
+### Bug Fixes
+
+* **stryker-api:** Update stryker-api peer dependency version ([677fc28](https://github.com/stryker-mutator/stryker/commit/677fc28))
+
+
+
+
+<a name="0.15.1"></a>
+## [0.15.1](https://github.com/stryker-mutator/stryker/compare/stryker-mocha-runner@0.15.0...stryker-mocha-runner@0.15.1) (2018-11-29)
+
+
+
+
+**Note:** Version bump only for package stryker-mocha-runner
+
 <a name="0.15.0"></a>
 # [0.15.0](https://github.com/stryker-mutator/stryker/compare/stryker-mocha-runner@0.14.6...stryker-mocha-runner@0.15.0) (2018-11-13)